feat: Add admin-created request flag
Introduces a new flag `createdByAdmin` to `LeaveRequest` and `User` types.
This flag allows tracking requests that are created by an administrator on behalf of a user.

The `createRequest` service method has been updated to accept an optional `targetUserId` and correctly set `createdByAdmin` based on whether the request is for the current user or a different user.

The dashboard now visually indicates requests created by administrators, enhancing transparency.